Patch for RENAME attribute on file opening (including minor refactoring for reload tests)
